update for 0.913
git-svn-id: https://path-of-exile-skilltree-planer.googlecode.com/svn/trunk@23 860b88a0-a686-10f0-e525-4207949ec8bc

+++ WPFSKillTree/WPFSKillTree/MainWindow.xaml.cs
@@ -110,8 +110,8 @@ namespace POESKillTree
             }
 
         }
-        private List<int> prePath;
-        private HashSet<int> toRemove;
+        private List<ushort> prePath;
+        private HashSet<ushort> toRemove;
         private bool justLoaded = false;
         private void comboBox1_SelectionChanged( object sender, SelectionChangedEventArgs e )
         {
@@ -158,7 +158,7 @@ namespace POESKillTree
             }
             if ( prePath != null )
             {
-                foreach ( int i in prePath )
+                foreach (ushort i in prePath)
                 {
                     Tree.SkilledNodes.Add( i );
                 }
